Remove CommConnectCaseSQL as a separate class
