+++ corehq/form_processor/models.py
@@ -18,7 +18,7 @@ from jsonobject import StringProperty
 from jsonobject.properties import BooleanProperty
 from lxml import etree
 from uuidfield import UUIDField
-
+from corehq.apps.sms.mixin import MessagingCaseContactMixin
 from corehq.blobs import get_blob_db
 from corehq.blobs.exceptions import NotFound, BadName
 from corehq.form_processor import signals
@@ -497,7 +497,7 @@ class SupplyPointCaseMixin(object):
 
 class CommCareCaseSQL(DisabledDbMixin, models.Model, RedisLockableMixIn,
                       AttachmentMixin, AbstractCommCareCase, TrackRelatedChanges,
-                      SupplyPointCaseMixin):
+                      SupplyPointCaseMixin, MessagingCaseContactMixin):
     objects = RestrictedManager()
 
     case_id = models.CharField(max_length=255, unique=True, db_index=True)
